seçao 16-17

+++ S14_ConsumeCart.NET6/GeekShopping/GeekShopping.CartAPI/Data/ValueObjects/ProductVO.cs
@@ -1,17 +1,12 @@
-﻿namespace GeekShopping.CartAPI.Data.ValueObjects {
-
-    public class ProductVO {
-        
+﻿namespace GeekShopping.CartAPI.Data.ValueObjects
+{
+    public class ProductVO
+    {
         public long Id { get; set; }
-
         public string Name { get; set; }
-       
-        public decimal Price { get; set; }
-       
+        public double Price { get; set; }
         public string Description { get; set; }
-       
         public string CategoryName { get; set; }
-        
         public string ImageURL { get; set; }
     }
 }
